+++ tests/test_doc__index_word_count.py
@@ -68,7 +68,7 @@ def test_doc__index_word_count():
     take_top_n = (
         c.this()
         .call_method("items")
-        .pipe(sorted, key=lambda t: t[1], reverse=True)
+        .sort(key=lambda t: t[1], reverse=True)
         .pipe(c.this()[: c.input_arg("top_n")])
         .as_type(dict)
     )